requirements.txt
pyaudio>=0.2.14
python-dotenv>=1.0.0
numpy>=1.24.0
websockets>=12.0
opencv-python>=4.8.0
pygame>=2.5.0
pyopengl>=3.1.0
glfw>=2.6.0
paho-mqtt>=1.6.0
scipy>=1.11.0
pedalboard>=0.9.0
